Remove extra props
